telegram_bot/requirements.txt
aiogram==2.25.1
asyncpg==0.27.0
python-dotenv==1.0.1
httpx==0.24.1
